Revision de issues a finalizar

+++ Codapli-Lemac-Arduino/adquisidor/lemac/Lvdt.h
@@ -2,7 +2,7 @@
 
 class Lvdt {
   private:
-    long tipo;
+    float tipo;
     int pinAnalog;
 
   public:
@@ -11,6 +11,6 @@ class Lvdt {
     Lvdt(int, int);
     void setTipo(int);
     int getTipo();
-    long getValue();
+    float getValue();
 
 };
